Fix benchmark.sh for the new cmake world.
It relied on reading gamedesc.txt to find a list of puzzle binaries to
run. But gamedesc.txt is now specific to the Windows build (since it
contains Windows executable names), and isn't available in the Unix
cmake build directory.

Fixed by making a simpler gamelist.txt available on all platforms.