gtk.c: squelch uninitialised-variable warning.
Apparently some compilers can't work out that new_window() will always
write to its error-message parameter if it returns failure, so they
complain at the call site that 'error' might be used uninitialised.

Fix by explicitly initialising it. (To NULL, which really _shouldn't_
stop the compiler from warning, because surely that's just as bad if
it reaches the following printf!)

Also, while I'm at it, move it into the block where it's used, so it
doesn't look as if it might pervade the whole of main().