tree234: avoid an uninitialised-variable warning.
Apparently, some compilers can't work out that the pattern

  if (!t->root) { special-case handler followed by early return }
  n = t->root;
  while (n) { ... }

will execute the while loop at least once, on the grounds that the
_first_ test for n being non-NULL must pass, because we initialised n
from t->root which can't be NULL on any code path where we didn't take
the early return. So they might give an uninitialised-variable warning
for the variable 'ki', which is initialised inside the while loop.

Compilers, eh. But it's easy enough to turn the while into a do-while,
so that even the least alert compiler will know it runs at least once.