gtk: Do not override window background colour when using a dark theme
When the user chooses a global dark theme, the window background will
be dark and the default text colour light.  Overriding the window
background to be light grey can make the menu and status bars
unreadable.  In case a dark theme is used, only set the background
colour for the content area.

References: https://bugs.debian.org/944237