Remove the concept of edge flags
You must hit the key itself. Determining whether we're really on an edge
is fiddly, especially in a multi-window world. This is certainly an
improvement over the previous state, which would too often have a massive
touch target for the leftmost key even when it was nowhere near the actual
edge of the screen.

Fixes #376