Load games on UI thread, only use a thread for process-watching
Loading a game from a saved string is near instantaneous, and doing it on
the UI thread avoids a lot of potential paths to crashes where the game
wants to manipulate a UI that has perhaps just gone away.

The only remaining thread-spawning is when we launch the game generation
process and need a thread to watch it and eventually read from it; that
thread's communication surface is thankfully quite small.