Narrow the intent-filter slightly
Turns out there are a lot of content:// intents exchanged in Android.
The case I first noticed was clicking on a calendar widget to open a
calendar event. Puzzles should not show up there! Most such intents
have a mime type that is not "application/octet-stream" so this filter
should now stay out of their way.