requestKeys: don't call it early; handle a possible null
The early call has always been more trouble than it's worth and is seemingly
the cause of almost all the remaining crashes in the current release.

I can't really see *how* my_game could be null, but that seems the only obvious
cause of a SIGSEGV there, so if it is, fail more gracefully.